Removed the QuantifiableMatch.java interface, created Matcher Interface, renamed the abstract Matcher class to AbstractMatcher, refactored code to account for name changes.
