+++ src/main/java/Matchers/Matcher.java
@@ -1,30 +1,27 @@
 package Matchers;
-import Entities.Entity;
 import Entities.Recipe;
-import Storages.Implementations.RecipeStorageImpl;
 
-import java.util.ArrayList;
 import java.util.List;
 
-// Matches ingredients with recipes
-public abstract class Matcher {
+public interface Matcher {
     /** Returns whether or not the specified entity contains all of the tags.
      *
      * @param recipe
      * @return whether or not the entity has all the tags..
      */
-    public abstract boolean matches(Recipe recipe);
+    boolean matches(Recipe recipe);
+
+    /**
+     * Return the match percentage between a criterion and the recipe in [0, 1]
+     * @param recipe
+     * @return Match percentage as double
+     */
+    double floatMatch (Recipe recipe);
 
     /**
      * Returns list of recipes for which you have ingredients
      * @param recipes is a list of recipes
      * @return list of recipes for which ingredients are had
      */
-    public List<Recipe> allMatches(List<Recipe> recipes) {
-        List<Recipe> matchedRecipes = new ArrayList<>();
-        for (Recipe recipe : recipes)
-            if (matches(recipe))
-                matchedRecipes.add(recipe);
-        return matchedRecipes;
-    }
+    List<Recipe> allMatches(List<Recipe> recipes);
 }
