--- src/main/java/Matchers/QuantifiableMatch.java
@@ -1,7 +1,0 @@
-package Matchers;
-
-import Entities.Recipe;
-
-public interface QuantifiableMatch {
-    double floatMatch (Recipe recipe);
-}
